Bugfix for issue #85 reported by Colin Nichol.

+++ CMN_binViewer.py
@@ -72,7 +72,7 @@ from module_highlightMeteorPath import highlightMeteorPath
 from module_CAMS2CMN import convert_rmsftp_to_cams
 from makeMP4 import makeMP4
 
-version = "3.37.1"
+version = "3.37.2"
 
 # set to true to disable the video radiobutton
 disable_UI_video = False
@@ -3377,24 +3377,30 @@ class BinViewer(Frame):
 
             # write filtered UFO-compatible R91 csv file 
             if sys.version_info[0] > 2:
-                self.timestamp_label.configure(text = "Updating UFO file...")
-                try:
-                    ufoFile = glob.glob(os.path.join(self.dir_path, '*.csv'))[0]
-                    
-                    with open(ufoFile,'r') as uf:
-                        ufoData = uf.readlines()
-                    newufoData = self.updateUFOData(FTPdetectinfoExport, ufoData)
-                    try:
-                        _, ufof = os.path.split(ufoFile)
-                        fnam = os.path.join(self.ConfirmationInstance.confirmationDirectory, ufof)
-                        with open(fnam, 'w') as newUfoFile:
-                            for line in newufoData:
-                                newUfoFile.write(line)
+               
+                ufoName = os.path.split(self.dir_path)[1] + '.csv'
+                ufoFile = os.path.join(self.dir_path, ufoName)
 
-                    except OSError as error:
-                        log.info('unable to write CSV file, {}'.format(error))
-                except FileNotFoundError:
-                    log.info('CSV file not present')
+                if os.path.isfile(ufoFile):
+                    self.timestamp_label.configure(text = "Updating UFO file...")
+                    try:
+                        with open(ufoFile,'r') as uf:
+                            ufoData = uf.readlines()
+                        newufoData = self.updateUFOData(FTPdetectinfoExport, ufoData)
+                        try:
+                            _, ufof = os.path.split(ufoFile)
+                            fnam = os.path.join(self.ConfirmationInstance.confirmationDirectory, ufof)
+                            with open(fnam, 'w') as newUfoFile:
+                                for line in newufoData:
+                                    newUfoFile.write(line)
+
+                        except OSError as error:
+                            log.info('unable to write CSV file, {}'.format(error))
+                    except Exception:
+                        log.info('UFO CSV file corrupt and cant be updated')
+                else:
+                    log.info('UFO CSV file not present')
+
             else:
                 print('ufo filtering doesnt work on Python 2.7')
